+++ Lab2/src/main/java/Labs_OOP_sem_3/service/PointService.java
@@ -36,11 +36,13 @@ public class PointService {
     public PointEntity update(PointDto pointDto) {
         FunctionEntity func = functionService.readByName(pointDto.getFunction().getName());
         PointEntity point = repository.findByFunctionIdAndPoint(func.getId(), pointDto.getX());
-        point.setYValue(pointDto.getY());
-        repository.save(point);
-        FunctionDto functionDto = ConvertToFuncDto.convert(func);
-        functionDto.setPoints(repository.findByFunction(functionDto.getId()));
-        functionService.update(functionDto);
+        if (point != null) {
+            point.setYValue(pointDto.getY());
+            repository.save(point);
+            FunctionDto functionDto = ConvertToFuncDto.convert(func);
+            functionDto.setPoints(repository.findByFunction(functionDto.getId()));
+            functionService.update(functionDto);
+        }
         return point;
     }
 
@@ -49,7 +51,9 @@ public class PointService {
     public void delete(PointDto pointDto) {
         FunctionEntity func = functionService.readByName(pointDto.getFunction().getName());
         PointEntity point = repository.findByFunctionIdAndPoint(func.getId(), pointDto.getX());
-        repository.deleteById(point.getId());
+        if (point.getId() != null) {
+            repository.deleteById(point.getId());
+        }
         FunctionDto functionDto = ConvertToFuncDto.convert(func);
         functionDto.setPoints(repository.findByFunction(functionDto.getId()));
         functionService.update(functionDto);
